+++ src/components/BlankStates/index.ts
@@ -1,13 +1,5 @@
-import WalletAssets from './WalletAssets/WalletAssets';
-import Save from './Save/Save';
-import Withdraw from './Withdraw/Withdraw';
-import Deposit from './Deposit/Deposit';
-import Exchange from './Exchange/Exchange';
+import { screens } from './screens';
 
 export default {
-	WalletAssets,
-	Save,
-	Withdraw,
-	Deposit,
-	Exchange
+	...screens
 };
