+++ src/components/BlankStates/screens/Deposit/Deposit.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import Box from '../Box/Box';
-import SettingsHeader from '../../SettingsHeader/SettingsHeader';
-import BlankLayout from '../BlankLayout/BlankLayout';
+import Box from '../../Box/Box';
+import SettingsHeader from '../../../SettingsHeader/SettingsHeader';
+import BlankLayout from '../../BlankLayout/BlankLayout';
 
-const Save = () => (
+const Deposit = () => (
 	<BlankLayout>
 		<SettingsHeader title="Wallet" onPress={() => null} />
 		<Box mt={136} mb={16} w={105} h={14} br={20} />
@@ -12,4 +12,4 @@ const Save = () => (
 	</BlankLayout>
 );
 
-export default Save;
+export default Deposit;
