+++ src/components/BlankStates/screens/index.ts
@@ -1,0 +1,13 @@
+import Deposit from './Deposit/Deposit';
+import Exchange from './Exchange/Exchange';
+import Save from './Save/Save';
+import WalletAssets from './WalletAssets/WalletAssets';
+import Withdraw from './Withdraw/Withdraw';
+
+export const screens = {
+	Deposit,
+	Exchange,
+	Save,
+	WalletAssets,
+	Withdraw
+};
